Support apps the must be closed before launch

+++ mine/manager.py
@@ -2,6 +2,8 @@
 
 import os
 import abc
+import time
+import glob
 import platform
 import functools
 
@@ -37,6 +39,19 @@ def log_running(func):  # pragma: no cover (manual)
 
 
 # TODO: enable coverage when a Linux test is implemented
+def log_starting(func):  # pragma: no cover (manual)
+    """Decorator for methods that start an application."""
+    @functools.wraps(func)
+    def wrapped(self, application):
+        """Wrapped method to log that an application is being started."""
+        log.info("starting %s...", application)
+        result = func(self, application)
+        log.info("running: %s", application)
+        return result
+    return wrapped
+
+
+# TODO: enable coverage when a Linux test is implemented
 def log_stopping(func):  # pragma: no cover (manual)
     """Decorator for methods that stop an application."""
     @functools.wraps(func)
@@ -63,12 +78,10 @@ class BaseManager(metaclass=abc.ABCMeta):  # pragma: no cover (abstract)
         """Determine if an application is currently running."""
         raise NotImplementedError
 
-# TODO: add this method when a feature calls for it
-# https://github.com/jacebrowning/mine/issues/5
-#     @abc.abstractclassmethod
-#     def start(self, application):
-#         """Start an application on the current computer."""
-#         raise NotImplementedError
+    @abc.abstractmethod
+    def start(self, application):
+        """Start an application on the current computer."""
+        raise NotImplementedError
 
     @abc.abstractmethod
     def stop(self, application):
@@ -89,6 +102,9 @@ class LinuxManager(BaseManager):  # pragma: no cover (manual)
             return None
         process = self._get_process(name)
         return process is not None
+
+    def start(self, application):
+        pass
 
     def stop(self, application):
         name = application.versions.linux
@@ -127,20 +143,39 @@ class MacManager(BaseManager):  # pragma: no cover (manual)
         process = self._get_process(name)
         return process is not None
 
+    @log_starting
+    def start(self, application):
+        name = application.versions.mac
+        if os.path.exists(name):
+            path = name
+        else:
+            path = os.path.join('/Applications', name)
+            if not os.path.exists(path):
+                pattern = os.path.join("/Applications/*", name)
+                log.debug("glob pattern: %s", pattern)
+                paths = glob.glob(pattern)
+                for path in paths:
+                    log.debug("match: %s", path)
+                assert paths, "not found: {}".format(application)
+                path = paths[0]
+        self._start_app(path)
+
     @log_stopping
     def stop(self, application):
         name = application.versions.mac
         process = self._get_process(name)
-        if process.is_running():
+        if process and process.is_running():
             process.terminate()
+            time.sleep(0.1)
 
     @staticmethod
     def _get_process(name):
         """Get a process whose executable path contains an app name."""
+        log.debug("searching for exe path containing '%s'...", name)
         for process in psutil.process_iter():
             try:
-                path = process.exe()
-                if name in path.split(os.sep):
+                path = process.exe().lower()
+                if name.lower() in path.split(os.sep):
                     if process.status() == psutil.STATUS_ZOMBIE:
                         log.debug("skipped zombie process: %s", path)
                     else:
@@ -148,6 +183,14 @@ class MacManager(BaseManager):  # pragma: no cover (manual)
                         return process
             except psutil.AccessDenied:
                 pass  # the process is likely owned by root
+
+    @staticmethod
+    def _start_app(path):
+        """Start an application from it's .app directory."""
+        assert os.path.exists(path), path
+        process = psutil.Popen(['open', path])
+        time.sleep(0.1)
+        return process
 
 
 class WindowsManager(BaseManager):  # pragma: no cover (manual)
@@ -158,6 +201,9 @@ class WindowsManager(BaseManager):  # pragma: no cover (manual)
     FRIENDLY = NAME
 
     def is_running(self, application):
+        pass
+
+    def start(self, application):
         pass
 
     def stop(self, application):
